+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - (#190) Possibility to `test` or `deploy` a service from a git or GitHub url
 - (#233) Add logs in the `service test` command with service logs by default and all dependencies logs with the `--full-logs` flag
 - (#235) Add `ListServices` and `GetService` APIs
+- (#) Add .mesgignore to excluding file from the Docker build
 
 #### Removed
 - (#234) Remove command `service status` in favor of `service list` command that includes status
@@ -30,5 +31,6 @@
 #### Fixed
 - (#179) [Doc] Outdated documentation for the CLI
 - (#185) Fix logs with extra characters when `mesg-core logs`
+- (#) Ignore files during Docker build
 
 ## [Unreleased]